+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog16
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 ## [Unreleased]
+
+- Added support for --without-api-ssl-verify to turn off SSL verification.
 
 ## [0.13.0] - 2019-08-20